Refactor taps.filter structure
