+++ taps/filter/__init__.py
@@ -1,13 +1,14 @@
 from __future__ import annotations
 
-from taps.filter.config import AllFilterConfig
-from taps.filter.config import FilterConfig
-from taps.filter.config import NullFilterConfig
-from taps.filter.config import ObjectSizeConfig
-from taps.filter.config import PickleSizeConfig
-from taps.filter.filters import AllFilter
-from taps.filter.filters import Filter
-from taps.filter.filters import NullFilter
-from taps.filter.filters import ObjectSizeFilter
-from taps.filter.filters import ObjectTypeFilter
-from taps.filter.filters import PickleSizeFilter
+from taps.filter.object import ObjectSizeFilter
+from taps.filter.object import ObjectSizeFilterConfig
+from taps.filter.object import ObjectTypeFilter
+from taps.filter.object import ObjectTypeFilterConfig
+from taps.filter.object import PickleSizeFilter
+from taps.filter.object import PickleSizeFilterConfig
+from taps.filter.protocol import Filter
+from taps.filter.protocol import FilterConfig
+from taps.filter.simple import AllFilter
+from taps.filter.simple import AllFilterConfig
+from taps.filter.simple import NullFilter
+from taps.filter.simple import NullFilterConfig
